requirements.txt
openai>=1.40.0,<2.0.0
pandas>=2.3.2,<3.0.0
numpy>=2.1.2,<3.0.0
python-dotenv>=1.0.1,<2.0.0
jinja2>=3.1.4,<4.0.0
selenium>=4.22.0,<5.0.0
rich>=13.7.1,<14.0.0
click>=8.1.7,<9.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
pydantic-settings>=2.5.2,<3.0.0
pytest>=8.3.3,<9.0.0
streamlit>=1.39.0,<2.0.0
sendgrid>=6.11.0,<7.0.0
